feat: refactor setup to use strategy pattern with legacy migration
Moves setup logic to internal/setup module and refactors to use new
strategy-based installation.

Key changes:
- InstallHook() now uses SelectInstallStrategy() to pick best approach
- Automatic migration from legacy inline hooks to external files
- HasLegacyInstall() detects old marker-based installations
- MigrateLegacyInstall() removes old hook and installs new one
- Preserves all user content during migration

Migration behavior:
- On first 'dirvana setup' after upgrade: auto-migrates legacy installs
- Removes 50+ lines of inline hook code from RC file
- Replaces with single source line to external hook file
- Shows clear migration message to user
